Allow requiring only specific driver support
