+++ lib/webdrivers/common.rb
@@ -2,8 +2,27 @@
 
 require 'rubygems/package'
 require 'zip'
+require 'webdrivers/logger'
+require 'selenium-webdriver'
 
 module Webdrivers
+  class << self
+    attr_accessor :proxy_addr, :proxy_port, :proxy_user, :proxy_pass, :install_dir
+
+    def logger
+      @logger ||= Webdrivers::Logger.new
+    end
+
+    def configure
+      yield self
+    end
+
+    def net_http_ssl_fix
+      raise 'Webdrivers.net_http_ssl_fix is no longer available.' \
+      ' Please see https://github.com/titusfortner/webdrivers#ssl_connect-errors.'
+    end
+  end
+
   class Common
     class << self
       attr_accessor :version
